Added a 10% risk budget
Many people cannot avoid COVID risks due to their work or personal
circumstances. Without this budget option the tool is less useful for
them as typical activities in their lives would all be just fully "red".
With this option they'd still have the chance to compare these choices.
Also, a 10% chance over a year will be much less in practice as we can expect
at-risk population to be vaccinated in the coming months.