Move prebuilt activities to the person-risk card
It was previously below the call to action. Since location needs to be
entered before the person-risk card appears, it was difficult to see
exactly what the old dropdown was doing. In addition, a handful of folks
who answered the survey and at least one user tester completely missed
seeing the prebuilts the first time around.

Make some related UX tweaks too:
* Adjust prebuilt activities box instructions so that the most important
  words for understanding are clustered at the beginning of the phrase.
  This also helps on mobile, where the prompt gets cut off.
* Add magnifying glass affordance to prebuilt activity search to more
  strongly indicate that it's for typing. It should probably be made
  clickable in the future.
* Fill country/state/county/preset activity on <enter> if there is only
  one possible item.