Sort locations alphabetically in the selector
With the new typeahead entry field it's less important for US states to
be at the top and it makes it less evident that the location names are
translated and should be typed according to the current locale. It is
also a bit disturbing if the list is not alphabetically ordered
according to the current locale.