i18n: localized country selector (#675)
* i18n: localized country selector (#671)

* Sort locations alphabetically in the selector

With the new typeahead entry field it's less important for US states to
be at the top and it makes it less evident that the location names are
translated and should be typed according to the current locale. It is
also a bit disturbing if the list is not alphabetically ordered
according to the current locale.

* Keep original order of locations for English

This way US states remain on top which makes them easier to select for
US users.