Update single shot efficacy of Moderna/Pfizer
The CDC reports an efficacy of 80% after 14 days after a first shot of Moderna or Pfizer, this figure includes both symptomatic and asymptomatic cases:
https://www.cdc.gov/mmwr/volumes/70/wr/mm7013e3.htm?s_cid=mm7013e3_w

When more data was collected, the figure was updated to 82%: https://www.cdc.gov/media/releases/2021/p0514-covid-19-vaccine-effectiveness.html

To me, based on these results, it would seem more reasonable to assume that asymptomatic cases are prevented at the same rate that symptomatic cases are prevented. If so, a pessimistic correction for asymptomatic as used in the current figure ((0.4 + .42 * .23) / (0.8 + .42 * .2) = 0.56) could be omitted.

Discussion:
- this data was gathered among health care workers only (although they did partial corrections for this)
- the 95% CI reported by CDC is 80% [59%–90%]
- the CDC warns that sensitivity of the used PCR tests could affect the outcome of the study

Is 56% still the best estimate? Is 20% more reasonable? Should we go for a middle ground? Let's discuss :D

p.s. I love the microcovid project, thanks for working on this!!!