Make update_prevalence more robust to negative daily cases (#860)
The JHU dataset is not strictly increasing case numbers (perhaps due to governments
correcting their numbers?). If a correction happens, compute # of cases in a week as
max(cumulative_cases_past_week) - min(cumulative_cases_past_week)

It is difficult to know if this is over or underestimating the actual number of cases,
since we have no way to determine the relative magnitudes of the correction vs the #
of new cases.