Nerfed speaking/shouting to assume only half-time
The old formula was modeling someone who was talking nonstop. I changed the formulas to model someone who's speaking / shouting half the time and silent half the time (roughly reasonable if we assume that every person is in a 1:1 conversation and biased high if we assume they're in groups of >2). Discussion at https://microcovid.slack.com/archives/C019MJ8RK45/p1635856803001400