Improve CSV-parsing typesafety by not assuming zero values
update_prevalence.py's current CSV parsing code assumes that a CSV
data element with no value (think the second entry in
`foo,,bar,1,baz`) maps to a 0 if we defined field as a non-optional
float or int.

I'd prefer we treat these as validation errors, which forces us to use
Optional[int]/Optional[float], which mypy will force us to handle the
None case when the value is returned.

I hope that this will alert us earlier if there are assumptions about
upstream data that don't hold for some data or have been violated due
to upstream changes.