Decouple Canadian vaccination data date ranges from case data ranges
It turns out that it's only by coincidence that we gather the same
dates today.  This change doesn't change the behavior of what we
fetch, just the structure of the code and comments to make that clear.

This is a refactor towards being more flexible on which case dates to
gather.